requirements.txt
# requirements.txt for the main mcp_operator project
datetime
vllm==0.7.3
# Core 
kerykeion==4.26.2


mistral_common
# Astrology Tool Dependencies
immanuel==1.4.4

pyswisseph==2.10.3.2 # This is typically installed as 'swisseph' via pip, though the import is 'swisseph'
svglib
# Image Search Tool Dependencies
requests

modelcontextprotocol[cli]
# Orpheus TTS Tool Dependencies
# The orpheus-speech library includes its own dependencies like torch and vllm.
orpheus-speech==0.1.0
pillow>=10.0.0
# You might need to explicitly add torch if orpheus-speech doesn't pull a compatible version
# or if you need a specific version for other reasons.
# torch # Add if needed, specify version if conflicts arise (e.g., torch==2.2.0)
transformers
# As per Orpheus docs, pinning vllm might be necessary due to bugs:
# vllm==0.7.3 # Uncomment if you encounter issues with the latest vllm from orpheus-speech
datasets

wandb 
accelerate 
trl 
# Add any other core dependencies your mcp_operator project needs
# (e.g., logging libraries, configuration parsing libraries, etc.)
# example_core_lib
google-api-python-client==2.109.0
google-auth-oauthlib
mcp
